separate avatar mask texture controller class
