+++ VMagicMirror/Assets/Baku/VMagicMirror/Scripts/Environment/AvatarMaskTextureController.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 92107e35783cee141859f42d826cef24
