+++ VMagicMirror/Assets/Baku/VMagicMirror/Scripts/Installer/EnvironmentInstaller.cs
@@ -24,6 +24,7 @@ namespace Baku.VMagicMirror.Installer
             container.BindInterfacesTo<ImageQualitySettingReceiver>().AsSingle();
             container.BindInterfacesAndSelfTo<CropAndOutlineController>().AsSingle();
             container.BindInterfacesAndSelfTo<CameraBackgroundColorController>().AsSingle();
+            container.BindInterfacesAndSelfTo<AvatarMaskTextureController>().AsSingle();
             
             // NOTE: サブキャラに依存している
             container.BindInterfacesAndSelfTo<Buddy.BuddyObjectRaycastChecker>().AsSingle();
